solo falta la funcionalidad de solicitar codigo de recuperacion

+++ src/main/java/co/edu/uniquindio/application/dtos/alojamiento/MetricasDTO.java
@@ -3,6 +3,6 @@ package co.edu.uniquindio.application.dtos.alojamiento;
 public record MetricasDTO(
         Integer totalResenas,
         Double promedioCalificaciones,
-        int totalReservas
+        Long totalReservas
 ) {
 }
